Create UI and ability to add/edit/remove projects
Projects can be added, edited, and removed from view.

New projects are added to the top of the list on the page and clicking
from project to project in the list will switch focus from one item to
another. This will be important when functionality for todos is added,
as the todos panel will be populated from the currently selected
project's todos list.