Fix a bug involving the last todo being focused
Fix a bug where focus automatically switches from the first todo in the
todo list to the last one right after adding a new todo, when either
adding a new project or focusing on a project in project list.

Using .push() to add a new todo to the project todo list meant new todos
were being added to the end of todoList[] on the project object.

Even though new todos were being added to the beginning of the list in
the DOM when they were created, when a new project was added or a
project was focused, the todos repopulated in the DOM according to how
they were stored in todoList[] on the project object.

Using .unshift() to match the behavior of how new todos were being added
to the DOM (add new todos to the beginning of todoList[]) fixes this
issue.